rename balancer::balancer to balancer::accept_http
