+++ src/balancer/mod.rs
@@ -1,4 +1,4 @@
-pub mod balancer;
+pub mod accept_http;
 pub mod format;
 mod response_errors;
 pub mod selection;
